CHG: appeased clippy
In a print statement you can add braches into the string and specify the
variables later via ', <variable name>' and that is how it was. Clippy
prefers it when the variable is inside the string '"... {<variable name>} ..."'.
A couple other things in a loop where brought in line with clippy.