NEW: added generation test
After discovering that a change meant to appease clippy broke the code
generation, it was decided to add a test to check for proper generation
of the qr code. For this purpose a test was added that generates a qr
code with the text "Hello, world! 123", error correction level M and no
debuggign output (debugging output irrelevant in the test). To detect
breaking changes early next time it happens.