FIX: rewrote the byte shuffling
Found a bug. If the program is called via
'cargo run -- "Exmaple string to test qr code programm for byte shuffeling" q'
is works, but called via
'cargo run -- "Example string to test qr code programm for byte shuffeling" q'
the program panics, because of an 'index out of bounds'-error.
All that changes is the word "Exmaple" is corrected to "Example", which
shouldn't have any influence on an index whatsoever.